test: refactor tests directory structure
Refactored directory structure of unit tests to mirror the Python
module structure. Moved E2E tests to `tests/e2e`.
