+++ tests/questionpy_sdk/commands/repo/test_index.py
